refactor(react-template): rename testWithCoverage utility and simplify logic
Refactor the testWithCoverage utility by renaming it to playwright.fixture.ts and simplifying the implementation. The new structure utilizes createTestWithCoverageFixture for better clarity and maintainability.
