+++ packages/react-template/tests/demo.test.ts
@@ -1,5 +1,5 @@
 import { expect } from '@playwright/test';
-import { describe, it } from './utils/playwright.testWithCoverage';
+import { describe, it } from './fixture/playwright.fixture';
 
 describe('Demo Test', () => {
   it('should have a title', async ({ page }) => {
